% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vegperiod.R
\name{calc_vegperiod}
\alias{calc_vegperiod}
\title{Calculate vegetation period start and end days of year)}
\usage{
calc_vegperiod(budburst.method, leaffall.method, dates = NULL,
  tavg = NULL, out.years = NULL, budburstdoy.fixed = 121,
  leaffalldoy.fixed = 279, ...)
}
\arguments{
\item{dates}{date vector passed to \code{\link{vegperiod}[vegperiod],
ignored if both leaffall.method and budburst.method = 'fixed'}}

\item{tavg}{vector or daily mean air temperature passed to \code{\link{vegperiod}[vegperiod],
ignored both if leaffall.method and budburst.method = 'fixed'}}

\item{out.years}{integer vector of the years to be returned. If not specified,
values for the years in dates will be returned.}

\item{budburstdoy.fixed}{vector of values to be returned if budburst.method = 'fixed'.}

\item{leaffalldoy.fixed}{vector of values to be returned if leaffall.method = 'fixed'.}
}
\value{
a data.frame with columns 'year', 'start', 'end'. if budburst.method = 'fixed' or leaffall.method = 'fixed',
'start' and 'end' contain the values specified in budburstdoy.fixed and leaffalldoy.fixed respectively.
}
\description{
Calculate vegetation period start and end days of year)
}
\examples{

calc_vegperiod(out.years = 2001:2010,
               budburst.method = "fixed",
               leaffall.method = "fixed",
               budburstdoy.fixed = floor(runif(10, 120,130)),
               leaffalldoy.fixed = floor(runif(2, 260,280)))
climate <- LWFBrook90R::slb1_meteo

calc_vegperiod(budburst.method = "Menzel",
               leaffall.method = "fixed",
               leaffalldoy.fixed = 280,
               dates = climate$dates,
               tavg = climate$tmean,
               species = "Fagus sylvatica",
               est.prev = 3)

calc_vegperiod(budburst.method = "Menzel",
               leaffall.method = "ETCCDI",
               dates = climate$dates,
               tavg = climate$tmean,
               species = "Quercus robur",
               est.prev = 3)
}
