% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_globrad.R
\name{CalcGlobRad}
\alias{CalcGlobRad}
\title{Calculate global solar radiation from sunshine duration hours}
\usage{
CalcGlobRad(doy, SunHour, lat, a0 = 0.25, b0 = 0.5)
}
\arguments{
\item{doy}{day of year}

\item{SunHour}{sunshine duration hours, same length as doy}

\item{lat}{latitude in decimal degrees}

\item{a0}{angström parameter a, defaults to 0.25}

\item{b0}{angström parameter b, defaults to 0.5}
}
\value{
a sequence of global radiation in \eqn{MJ m^-2 d^-1} with the length of doy
}
\description{
Uses day of year (doy) and latitude to determine extraterrestrial solar radiation and daylength,
which is then used to calculate global radiation using angström formula
}
