% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_swatday.R
\name{extract_from_swatday.asc}
\alias{extract_from_swatday.asc}
\title{Extract variables from swatday.asc data.tables and organize layer-wise values in columns}
\usage{
extract_from_swatday.asc(dat, layers = NULL, vars = NULL)
}
\arguments{
\item{dat}{swatday.asc data.table as returned from \code{\link{readOutput.B90}}:}

\item{layers}{integer vector addressing the layer numbers (nl) to be extracted
from dat. If not supplied, values from all layers will be returned}

\item{vars}{character vector containing variable names (out of "swati","theta",
"wetnes","psimi","psiti" to be extracted from dat. If not supplied
all 5 variables will be returned.}
}
\value{
a data.table with the layers' daily values of the variables organized in columns
    (wide format) with the names being made up of the variable name and layer number.
}
\description{
Extract variables from swatday.asc data.tables and organize layer-wise values in columns
}
\examples{
# Read files from LWF-Brook90 output directory

output <- readOutput.B90("tests/out/")

# Reorganize all variables for all layers

swati_wide <- extract_from_swatday.asc(output$swatday.asc)

# select single variables and layers:
swati_wide <- extract_from_swatday.asc(output$swatday.asc,
                                       layers =1:10,
                                       vars = c("swat", "psim", "theta"))

}
