% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptf_hamken_ff.R
\name{hydpar_forestfloor_hamken}
\alias{hydpar_forestfloor_hamken}
\title{Mualem - van Genuchten Parameters for forest floor organic soil horizons, from Hammel & Kennel 2001}
\usage{
hydpar_forestfloor_hamken(n = 1)
}
\arguments{
\item{n}{An integer value specifying the number of rows the returned data.frame should have
(i.e. the number of repetitions of the MvG-Parameter set)}
}
\value{
A data.frame with the number of rows equal to length(tex.KA5)
The data.frame contains the following variables:
\describe{
  \item{ths}{Saturation water content fraction}
  \item{thr}{Residual water content fraction}
  \item{alpha}{Alpha parameter of van Genuchten water retention function, m-1 }
  \item{npar}{N parameter of van Genuchten water retention function}
  \item{mpar}{M parameter of van Genuchten water retention function}
  \item{ksat}{Saturated hyraulic conductivity parameter of Mualem hydraulic conductivity function, mm d-1}
  \item{tort}{Tortuosity parameter of Mualem hydraulic conductivity function}
}
}
\description{
Derives a set of values for MVG Parameters ThetaR, ThetaS, Alpha, n, Tort and Ksat
for forest floor horizons as decribed in Hammel & Kennel 2001.
}
\references{
Hammel & Kennel (2001) Charakterisierung und Analyse der Wasserverfügbarkeit
                und des Wasserhaushalts von Waldstandorten in Bayern mit dem Simulationsmodell BROOK90,
                Forstliche Forschungsberichte München 185
}
