% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptf_hypres_regr.R
\name{hydpar_hypres}
\alias{hydpar_hypres}
\title{Mualem - van Genuchten Parameters from regression functions developed by Woesten et al 1999}
\usage{
hydpar_hypres(clay, silt, bd, oc.pct = 0.1, topsoil = TRUE,
  humconv = 1.72)
}
\arguments{
\item{clay}{clay mass percent}

\item{silt}{silt mass percent}

\item{bd}{bulk density in g cm-3}

\item{oc.pct}{organic carbon content in mass percent}

\item{topsoil}{Is the sample from the topsoil?}

\item{humconv}{conversion factor oc.pct to organic matter percent}
}
\value{
A data.frame with the number of rows equal to length(clay)
The data.frame contains the following variables:
\describe{
  \item{ths}{Saturation water content fraction}
  \item{thr}{Residual water content fraction}
  \item{alpha}{Alpha parameter of van Genuchten water retention function, m-1}
  \item{npar}{N parameter of van Genuchten water retention function}
  \item{mpar}{M parameter of van Genuchten water retention function}
  \item{ksat}{Saturated hyraulic conductivity parameter of Mualem hydraulic conductivity function, mm d-1}
  \item{tort}{Tortuosity parameter of Mualem hydraulic conductivity function}
}
}
\description{
Derives the Mualem - van Genuchten parameters ThetaR, ThetaS, Alpha, n, Tort and Ksat
     from sand, silt, clay, bulk density, organic carbon content values
}
\examples{
hydpar_hypres(20,20,1.5,2)

}
\references{
Woesten JHM, Lilly A, Nemes A, Le Bas C  (1999) Development and use of
                a database of hydraulic properties of European soils. Geoderma 90, pp. 169-185
}
