% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptf_hypres_tab.R
\name{hydpar_hypres_tab}
\alias{hydpar_hypres_tab}
\title{Mualem - van Genuchten Parameters from W?sten et al. 1999, Tab. 4}
\usage{
hydpar_hypres_tab(tex.hypres, topsoil)
}
\arguments{
\item{tex.hypres}{A character vector with soil texture values according to classes applied in Woesten et al. 2009}

\item{topsoil}{logical Is the sample from the topsoil?}
}
\value{
A data.frame with the number of rows equal to length(tex.hypres)
The data.frame contains the following variables:
\describe{
  \item{ths}{Saturation water content fraction}
  \item{thr}{Residual water content fraction}
  \item{alpha}{Alpha parameter of van Genuchten water retention function, m-1}
  \item{npar}{N parameter of van Genuchten water retention function}
  \item{mpar}{M parameter of van Genuchten water retention function}
  \item{ksat}{Saturated hyraulic conductivity parameter of Mualem hydraulic conductivity function, mm d-1}
  \item{tort}{Tortuosity parameter of Mualem hydraulic conductivity function}
}
}
\description{
Derives the MVG Parameters ThetaR, ThetaS, Alpha, n, Tort and Ksat
    from soil texture classes and bulk density values
}
\examples{
hydpar_hypres_tab(tex.hypres = c("C","MF"), topsoil = c(TRUE,FALSE))

}
\references{
Woesten JHM, Lilly A, Nemes A, Le Bas C  (1999) Development and use of
                a database of hydraulic properties of European soils. Geoderma 90, pp. 169-185
}
