% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptf_puh2.R
\name{hydpar_puh2}
\alias{hydpar_puh2}
\title{Mualem - van Genuchten Parameters from regression functions developed by Puhlmann & von Wilpert 2011}
\usage{
hydpar_puh2(clay, silt, sand, bd, oc.pct = 0.5)
}
\arguments{
\item{clay}{clay mass percent}

\item{silt}{silt mass percent}

\item{sand}{sand mass percent}

\item{bd}{bulk density in g cm-3}

\item{oc.pct}{organic carbon content in mass percent}
}
\value{
A data.frame with the number of rows equal to length(tex.KA5)
The data.frame contains the following variables:
\describe{
  \item{ths}{Saturation water content fraction}
  \item{thr}{Residual water content fraction}
  \item{npar}{N parameter of van Genuchten water retention function}
  \item{mpar}{M parameter of van Genuchten water retention function}
  \item{alpha}{Alpha parameter of van Genuchten water retention function, m-1}
  \item{ksat}{Saturated hyraulic conductivity parameter of Mualem hydraulic conductivity function, mm d-1}
  \item{tort}{Tortuosity parameter of Mualem hydraulic conductivity function}
}
}
\description{
Derives the Mualem - van Genuchten parameters ThetaR, ThetaS, Alpha, n, Tort and Ksat
    from clay, silt, sand contents, bulk density, and organic carbon content values
}
\examples{
hydpar_puh2(20,20,60,1.5)
}
\references{
Puhlmann H, von Wilpert K (2011) Testing and development of pedotransfer
                functions for water retention and hydraulic conductivity of forest soils.
                Wald?kologie, Landschaftsforschung und Naturschutz 12, pp. 61-71
}
