% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptf_wessolek_mvg.R
\name{hydpar_wessolek_mvg}
\alias{hydpar_wessolek_mvg}
\title{Mualem - van Genuchten Parameters from Wessolek et al. 2009, Table 10.}
\usage{
hydpar_wessolek_mvg(tex.KA5)
}
\arguments{
\item{tex.KA5}{A character vector with soil texture values according to KA5}
}
\value{
A data.frame with the number of rows equal to length(tex.KA5)
The data.frame contains the following variables:
\describe{
  \item{tex.KA5}{soil texture according to KA5 }
  \item{thr}{Residual water content fraction}
  \item{ths}{Saturation water content fraction}
  \item{alpha}{Alpha parameter of van Genuchten water retention function, m-1}
  \item{npar}{N parameter of van Genuchten water retention parameter}
  \item{mpar}{M parameter of van Genuchten water retention parameter}
  \item{tort}{Tortuosity parameter of Mualem hydraulic conductivity function}
  \item{ksat}{Saturated hyraulic conductivity parameter of Mualem hydraulic conductivity function, mm d-1}
}
}
\description{
Derives the MVG Parameters ThetaR, ThetaS, Alpha, n, Tort and Ksat
    from KA5 soil texture classes
}
\references{
Wessolek, Kaupenjohann and Renger (2009) Bodenphysikalische Kennwerte
                und Berechnungsverfahren f?r die Praxis. Boden?kologie und Bodengenese 40,
                Berlin, Germany (Table No. 10)
}
