% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeRoots.R
\name{MakeRelRootDens}
\alias{MakeRelRootDens}
\title{Generates a root density depth function the soil layers' lower depth limits}
\usage{
MakeRelRootDens(soilnodes, maxrootdepth = min(soilnodes),
  method = "betamodel", beta = 0.97, relrootden = NULL,
  rootdepths = NULL)
}
\arguments{
\item{soilnodes}{vector of lower soil layer depth limits,
for which the relative root distribution will be calculated.}

\item{maxrootdepth}{the maximum rooting depth (m, negative downwards) below which
relative root length density will be set to zero}

\item{method}{method name for the root depth distribution. 'betamodel' uses the model after Gale & Grigal,
'table' interpolates the value pairs of 'rootdepths' and 'relrootden' to 'soilnodes'.
'linear' returns linearly decreasing root densities with the maximum value
(taken from the first vector element of 'relrootden') at the uppermost layer to 0
at 'maxrootdepth'. 'constant' returns a uniform root distribution with depth
corresponding to the first vector element of 'relrootden' and 0 below 'maxrootdepth'.}

\item{beta}{parameter(s) of the root distribution function}

\item{relrootden}{vector of relative root densities}

\item{rootdepths}{vector of lower depths limit, corresponding to 'relrootden'}
}
\value{
vector of relative rootlength, corresponding to soilnodes
}
\description{
Generates a root density depth function the soil layers' lower depth limits
}
\examples{

MakeRelRootDens(soilnodes = seq(0,-2, by = -0.01),
method = "betamodel",
beta = 0.93)

MakeRelRootDens(soilnodes = soil_slb1$lower,
maxrootdepth = -1.4,
method = "linear",
relrootden = 0.2)

MakeRelRootDens(soilnodes = soil_slb1$lower,
maxrootdepth = -1.4,
method = "table",
relrootden =  c(15, 35, 15, 7.5, 4, 12, 2, 2, 0),
rootdepths = c(-0.02, -0.15, -0.35, -0.5, -0.65,-0.9,-1.1,-1.3,-1.6))

}
