% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSeasLAI.R
\name{MakeSeasLAI}
\alias{MakeSeasLAI}
\title{Construct the seasonal course of leaf area index from parameters}
\usage{
MakeSeasLAI(method = "b90", year, maxlai, winlaifrac = 0,
  budburst.doy = 121, leaffall.doy = 279, emerge.dur = 28,
  leaffall.dur = 58, shape.optdoy = 220, shape.budburst = 0.5,
  shape.leaffall = 10, lai.doy = c(1, 121, 150, 280, 320, 365),
  lai.frac = c(0, 0, 0.5, 1, 0.5, 0))
}
\arguments{
\item{method}{name of method for generating the sequence. Must be one of "b90", "linear", "Coupmodel"}

\item{year}{years to be returned}

\item{maxlai}{maximum value during summer}

\item{winlaifrac}{fraction of maxlai during winter (ignored when method = 'linear')}

\item{budburst.doy}{budburst day of year (ignored when method = 'linear')}

\item{leaffall.doy}{day of year when leaf fall begins (ignored when method = 'linear')}

\item{emerge.dur}{number of days from budburst until maximum leaf area index is reached}

\item{leaffall.dur}{number of days until minimum leaf are index is reached}

\item{shape.budburst}{shape parameter for the growth phase (required when method = "Coupmodel")}

\item{shape.leaffall}{shape parameter growth cessation (required when method = "Coupmodel")}

\item{lai.doy}{integer vector of days of years}

\item{lai.frac}{vector of values of fractional leaf area index corresponding
to lai.doy (required when method = "linear")}

\item{opt.doy}{day of year when optimum value is reached (required when method = "Coupmodel")}
}
\value{
a vector of daily lai values covering the years specified
}
\description{
A daily sequence of leaf area index is derived from maximum and minimum values,
dates and shape parameters using different methods.
}
\examples{
# leaf area index for a single year

param.b90 <- MakeParam.B90()

plot(MakeSeasLAI(method = "linear",
year = 2001,
maxlai = param.b90$maxlai,
winlaifrac = param.b90$winlaifrac,
budburst.doy = param.b90$budburstdoy,
leaffall.doy = param.b90$leaffalldoy,
emerge.dur = param.b90$emergedur,
leaffall.dur = param.b90$leaffalldur))

plot(MakeSeasLAI(method = "Coupmodel",
                 year = 2001:2003,
                 maxlai = param.b90$maxlai,
                 winlaifrac = param.b90$winlaifrac,
                 budburst.doy = param.b90$budburstdoy,
                 leaffall.doy = param.b90$leaffalldoy,
                 leaffall.dur = param.b90$leaffalldur,
                 shape.optdoy = 180,
                 shape.budburst = 0.5,
                 shape.leaffall = 5))

# leaf area index for multiple years, using multiple values
plot(MakeSeasLAI(method = "linear",
year = 2001:2003,
maxlai = c(6,4,5),
winlaifrac = param.b90$winlaifrac,
budburst.doy = param.b90$budburstdoy,
leaffall.doy = param.b90$leaffalldoy,
emerge.dur = param.b90$emergedur,
leaffall.dur = param.b90$leaffalldur))
}
