% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msiterunLWFB90.R
\name{msiterunLWFB90}
\alias{msiterunLWFB90}
\title{Make a multi-site simulation using list of inputs.}
\usage{
msiterunLWFB90(param.b90, options.b90, climate, soil = NULL,
  all_combinations = F, multirun.dir = "MultiRuns/",
  keep.subdirs = FALSE, cores = 2, showProgress = TRUE, ...)
}
\arguments{
\item{param.b90}{Named list of parameters to be used in all simulations, or a list of multiple parameter sets.}

\item{options.b90}{Named list of model control options to be used in all simulations, or a list of multiple lists with different options sets.}

\item{climate}{data.frame with climate data, or a list of climate data.frames}

\item{soil}{data.frame with soil properties to be used in all simulations, or a list of data.frames with different soil profiles}

\item{all_combinations}{logical Set up and run all possible combinations of
individual param.b90, options.b90, climate and soil objects? Default is FALSE,
running one or multiple param.b90/options.b90 combinations for a series of climate/soil combinations.}

\item{multirun.dir}{directory name where to create the subdirectories for the single runs. Default is 'MultiRuns/'.}

\item{keep.subdirs}{keep sub-directories of the single runs? Default is FALSE}

\item{cores}{number of cores to use for parallel processing}

\item{showProgress}{Show progressbar?}

\item{...}{Further arguments passed to \code{\link[runLWFB90]{runLWFB90}}.}
}
\value{
A named list with the results of the single runs as returned by \code{\link{runLWFB90}}.
Simulation or processing errors are passed on. The names of the returned list entries
are concatenated from the names of the input list entries
in the following form: <climate> <soil> <param.b90> <option.b90>.
}
\description{
Repeatedly calls \code{\link[runLWFB90]{runLWFB90}} for combinations of climate and soil data,
and parameters and model control options, and returns the singlerun-results as a named list.
}
\examples{
options.b90 <- setoptions_LWFB90(budburst.method = "Menzel")

# define parameter sets
param_l <- list(spruce = setparam_LWFB90(maxlai = 5, budburst.species = "Picea abies (frueh)", winlaifrac = 0.8),
                beech = setparam_LWFB90(maxlai = 6, budburst.species = "Fagus sylvatica", winlaifrac = 0))

soil <- cbind(slb1_soil, hydpar_wessolek_mvg(slb1_soil$texture))

# define list of soil objects
soils <- list(soil1 = soil, soil2 = soil)

# define list of climate objects
climates <- list(clim1 = slb1_meteo, clim2 = slb1_meteo)

# run two parameter sets on a series of climate and soil-objects
res <- msiterunLWFB90(param.b90 = param_l,
                      options.b90 = options.b90,
                      soil = soils,
                      climate = climates)
names(res)

# all combinations
res <- msiterunLWFB90(param.b90 = param_l,
                      options.b90 = options.b90,
                      soil = soils,
                      climate = climates,
                      all_combinations = T)
names(res)
}
