% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrun_helper_listlength.R
\name{nstlist_length}
\alias{nstlist_length}
\title{Determine list length of nested lists}
\usage{
nstlist_length(ll)
}
\arguments{
\item{ll}{a list}
}
\value{
integer value with the length at uppermost level. Flat lists have length of 1.
}
\description{
Determine list length of nested lists
}
