% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_to_rbrook90.R
\name{param_to_rbrook90}
\alias{param_to_rbrook90}
\title{Create a parameter vector for the r_lwfbrook90-function}
\usage{
param_to_rbrook90(param.b90, imodel)
}
\arguments{
\item{param.b90}{a named list of model parameters.}

\item{imodel}{name of hydraulic model ('MvG' or 'CH')}
}
\value{
a numerical vector with the parameters in the right order for r_brook90
}
\description{
The parameter vector for r_lwfbrook90 is created from the lists of model controls options, parameters, and soil.
}
