% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_to_rbrook90.R
\name{param_to_rlwfbrook90}
\alias{param_to_rlwfbrook90}
\title{Create a parameter vector for the r_lwfbrook90-function}
\usage{
param_to_rlwfbrook90(param_b90, imodel)
}
\arguments{
\item{param_b90}{A named list of model parameters.}

\item{imodel}{Name of hydraulic model ('MvG' or 'CH')}
}
\value{
A numerical vector with the parameters in the right order for
  \code{\link{r_lwfbrook90}}.
}
\description{
The \code{param} vector for \code{\link{r_lwfbrook90}} is created from model
parameters.
}
