% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_to_rbrook90.R
\name{parms_to_rbrook90}
\alias{parms_to_rbrook90}
\alias{param_to_rlwfbrook90}
\alias{siteparam_to_rlwfbrook90}
\title{Create parameter vectors for the LWF-Brook90 Fortran library}
\usage{
param_to_rlwfbrook90(b90f_input)

siteparam_to_rlwfbrook90(b90f_input)
}
\arguments{
\item{b90f_input}{A named list of model parameters.}
}
\value{
A numeric vector with the parameters in the required order.
}
\description{
Functions to create the \code{param} and \code{siteparam} vectors to be passed to the C-interface.
}
