% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantcoupmodel.R
\name{plant.coupmodel}
\alias{plant.coupmodel}
\title{Interpolate plant properties using the 'Coupmodel' method.}
\usage{
plant.coupmodel(minval, maxval, doy.incr, doy.max, doy.min, shape.incr,
  shape.decr, maxdoy)
}
\arguments{
\item{minval}{Minimum value.}

\item{maxval}{Maximum value.}

\item{doy.incr}{Date (day of year) when increasing to maxval begins.}

\item{doy.max}{Date (day of year) when maxval is reached.}

\item{doy.min}{Date (day of year) when minval is reached.}

\item{shape.incr}{Shape parameter of the increasing phase.}

\item{shape.decr}{Shape parameter of the decreasing phase.}

\item{maxdoy}{Length of the year, 366 for leap years, 365 for normal years.}
}
\value{
A numeric vector of length \code{maxdoy}.
}
\description{
Creates a daily sequence for one year from parameters
}
\examples{
plot(plant.coupmodel(0,5, 121, 200, 280, 0.3, 3, 365))
}
\references{
Jansson, P.-E. & Karlberg, L. (2004): "Coupled heat and mass transfer model for
soil-plant-atmosphere systems."
\emph{Royal Institute of Technolgy, Dept of Civil and Environmental Engineering Stockholm}
}
