% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantlinear.R
\name{plant.linear}
\alias{plant.linear}
\title{Interpolate plant properties using the 'linear' method.}
\usage{
plant.linear(doys, values, maxdoy)
}
\arguments{
\item{doys}{dates (days of year)}

\item{values}{values}

\item{maxdoy}{length of the year, 366 for leap years, 365 for normal years}
}
\value{
numeric vector of length maxdoy
}
\description{
Creates a daily sequence for one year from doy/value pairs.
}
\examples{
plot(plant.linear(c(110,200,250,280), c(0,0.8,1,0)), 365)
}
