% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richter_prec_corr.R
\name{prec_corr}
\alias{prec_corr}
\title{Correct rain gauge precipitation data for wind and evaporation errors after
Richter (1995)}
\usage{
prec_corr(month, tavg, prec, station.exposure = "mg", full.output = FALSE)
}
\arguments{
\item{month}{Vector of months.}

\item{tavg}{Vector of air temperature values (deg C). Same length as
\code{month}.}

\item{prec}{Vector of measured rainfall vales (mm). Same length as
\code{month}.}

\item{station.exposure}{Situation of the weather station where prec was
measured: one of 'frei', 'lg', 'mg', 'sg' (corresponding to full exposure,
low protected, moderate protected, strong protected situation).}

\item{full.output}{Logical wether to return the full data set additionally
including input data, correction coefficients.}
}
\value{
A vector of corrected rainfall data, or (if \code{full.output ==
  TRUE}) a data.table containing the input objects, the month, the
  precipitation type ('N4So': liquid rain, summer; 'N4Wi' liquid rain,
  winter; 'N8' = sleet, 'N7' = snow), correction coefficients epsilon and b,
  and the corrected rainfall.
}
\description{
Correct rain gauge precipitation data for wind and evaporation errors after
Richter (1995)
}
\examples{
clim <- slb1_meteo[as.integer(format(slb1_meteo$dates,"\%Y")) \%in\% 2001:2005,]
clim$month <- as.integer(format(clim$dates, "\%m"))

prec_meas <- clim$prec
prec_corr_frei <- with(clim,
                       prec_corr(month, tmean, prec, station.exposure = "frei"))
prec_corr_lg <- with(clim,
                     prec_corr(month, tmean, prec, station.exposure = "lg"))
prec_corr_mg <- with(clim,
                     prec_corr(month, tmean, prec, station.exposure = "mg"))
prec_corr_sg <- with(clim,
                     prec_corr(month, tmean, prec, station.exposure = "sg"))

plot(clim$dates, cumsum(prec_corr_frei),
type = "l", col = "violet", xlab = "dates", ylab = "cum. precipitation (mm)")
lines(clim$dates, cumsum(prec_corr_lg), col = "blue")
lines(clim$dates, cumsum(prec_corr_mg), col = "green")
lines(clim$dates, cumsum(prec_corr_sg), col = "red")
lines(clim$dates, cumsum(prec_meas))
legend('bottomright', c('frei', "lg", "mg", "sg"),
       col = c("violet", "blue", "green", "red", "black"),
       lty = 1, pch = NULL )
}
\references{
Richter, D. (1995)
Ergebnisse methodischer Untersuchungen zur Korrektur des systematischen
Messfehlers des Hellmann-Niederschlagsmessers.
\emph{Berichte des Deutschen Wetterdienstes}, \bold{194}, 93 pp, Offenbach,
Germany
}
