% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_brook90.R
\name{r_brook90}
\alias{r_brook90}
\title{Define the function to run 3PGN in R}
\usage{
r_brook90(site, climate, param, paramYear, materials, soil, pfile = NULL,
  output)
}
\arguments{
\item{site}{a one row matrix with site level information}

\item{climate}{a matrix with 15 columns of climatic data}

\item{param}{a vector of parameters input.}

\item{paramYear}{a matrix of the yearly input parameters. Year column corresponds to one month.}

\item{materials}{a matrix of the materials parameters}

\item{soil}{a matrix of soil parameters}

\item{pfile}{a precipitation file}

\item{output}{a matrix of output settings}
}
\description{
Define the function to run 3PGN in R
}
\details{
This is the model
}
