% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrun_helper_replace_vec_elem.R
\name{replace_vecelements}
\alias{replace_vecelements}
\title{Replace elements in a data.frame or vector of length > 1 by name}
\usage{
replace_vecelements(x, varnms, vals)
}
\arguments{
\item{x}{vector or dataframe}

\item{varnms}{variable names to match: Specify}

\item{vals}{values to insert}
}
\value{
The vector or data.frame in x with the elements 'varnms' replaced by vals.
}
\description{
Replace elements in a data.frame or vector of length > 1 by name
}
\examples{


x =param.b90[["soil_materials"]]
varnms = "soil_materials.ths1"
vals = 0.999
replace_vecelements(x, varnms, vals)
x <- param.b90[["pdur"]]
varnms <- c("pdur2", "pdur12")
vals <- c(0,10)
replace_vecelements(x, varnms, vals)
}
