% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLWFB90.R
\name{runLWFB90}
\alias{runLWFB90}
\title{Run the LWF-Brook90 hydrological model}
\usage{
runLWFB90(project.dir = "runLWFB90/", options.b90, param.b90, climate,
  precip = NULL, soil = NULL, output = setoutput_LWFB90(),
  obs = NULL, gof_fun = NULL, rtrn.input = TRUE,
  rtrn.output = TRUE, read.output = TRUE, chk.input = TRUE,
  output.log = TRUE, run = TRUE, verbose = TRUE)
}
\arguments{
\item{project.dir}{directory-name of the project to which output files
are written. Will be created, if not existing. Defaults to 'runLWFB90/'}

\item{options.b90}{named list of model control options. Use
\code{\link{setoptions_LWFB90}} to generate a list with default model control options.}

\item{param.b90}{Named list of model input parameters. Use
\code{\link{setparam_LWFB90}} to generate a list with default model parameters.}

\item{climate}{data.frame with daily climate data with columns named according to
 \emph{dates}, \emph{tmax}, \emph{tmin}, \emph{tmean}, \emph{wind}, \emph{prec}, \emph{vappres},
and either \emph{globrad} or \emph{sunhours}). When using 'sunhours', please set options.b90$fornetrad = 'sunhours'}

\item{precip}{data.frame with columns 'dates' and 'prec' to supply precipitation data separately from climate data.
Can be used to provide sub-day resolution precipitation data to LWFBrook90. For each day in dates,
1 (daily resolution) to 240 values of precipitation can be provided, with the number of values
per day defined in options.b90$prec.interval.}

\item{soil}{data.frame containing the hydraulic properties of the soil layers.
Each row represents one layer, containing the layers' boundaries and soil hydraulic parameters.
The column names for the upper and lower layer boundaries are \emph{upper} and \emph{lower} (m, negative downwards).
When using options.b90$imodel = 'MvG', the hydraulic parameters are  \emph{ths}, \emph{thr},
 \emph{alpha} [m-1], \emph{npar}, \emph{ksat} [mm d-1] and \emph{tort}.  With options.b90$imodel = 'CH',
 the parameters are \emph{thsat} , \emph{thetaf},\emph{psif} [kPa], \emph{bexp},
 \emph{kf} (mm d-1), and \emph{wetinf}. For both parameterizations, the volume fraction of stones has to be named \emph{gravel}.
 If the soil data.frame is not provided, list items 'soil_nodes' and 'soil_materials'
  of param.b90 are used for the simulation. These have to be set up in advance, see \code{\link{soil_to_param}}.}

\item{output}{a [10,5]-matrix flagging the desired model-output. Use
\code{\link{setoutput_LWFB90}} to generate and edit a default output matrix.}

\item{obs}{a data.frame with observations. The names of observations are looked up
in the list of simulation output data.frames, and goodness-of-fit measures are calculated for each
equivalant found. See \code{\link{calc_gof}}. Disabled when read.output = F.}

\item{gof_fun}{a function of form function(sim,obs) or list of functions passed to \code{\link{calc_gof}}.
Disabled when read.output = F.}

\item{rtrn.input}{append 'param.b90', 'options.b90', 'soil' and daily plant
properties ('standprop_daily', as derived from parameters) to the result?}

\item{rtrn.output}{return the simulation results? Disabled when read.output = F.}

\item{read.output}{read the simulation result files from project.dir? Default is TRUE.}

\item{chk.input}{logical wether to check param.b90, options.b90, climate, precip, soil, and obs
for completeness and consistency.}

\item{output.log}{write the logfile 'Log.txt' to the 'project.dir'? Default is TRUE.}

\item{run}{run LWF-Brook90 or only return model input objects?
Useful to inspect the effects of options and parameters on model input. Default is TRUE.}

\item{verbose}{print messages to the console? Default is TRUE.}
}
\value{
A list containing the contents of the LWF-Brook90 output files found in 'project.dir' along with model_input,
goodness-of-fit measures, and duration and finishing time of the simulation.
}
\description{
Sets up the input objects for the LWF-Brook90 hydrological model, starts the model,
and returns the results if desired.
}
\examples{
# Set up lists containing model control options and model parameters:
param.b90 <- setparam_LWFB90()
options.b90 <- setoptions_LWFB90()

# Set start and end Dates for the simulation
options.b90$startdate <- as.Date("2002-2-15")
options.b90$enddate <- as.Date("2005-07-5")

# Derive soil hydraulic properties from soil physical properties
# using pedotransfer functions
soil <- cbind(slb1_soil, hydpar_wessolek_mvg(slb1_soil$texture))

# Run LWF-Brook90
b90.result <- runLWFB90(project.dir = "example_run_b90",
                        options.b90 = options.b90,
                        param.b90 = param.b90,
                        climate = slb1_meteo,
                        soil = soil)

# use observations to calculate goodness of fit
observations <- slb1_mpot #'daily water potential in different soil depths
# prepare data: names have to be found in simulation output.
names(observations)[2:6] <- c("psimi5", "psimi7", "psimi10", "psimi16","psimi21")

# Fit-functions
gof <- list(nse = hydroGOF::NSE,
            me = function(sim,obs){mean(sim-obs, na.rm=T)})

# run model, but only return gof-function results
b90.gofmpot <- runLWFB90(project.dir = "example_run_b90",
                         options.b90 = options.b90,
                         param.b90 = param.b90,
                         climate = slb1_meteo,
                         soil = soil,
                         obs = observations,
                         gof_fun = gof,
                         rtrn.output = F,
                         rtrn.input = F)
b90.gofmpot

}
