% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msite_combinations.R
\name{setup_combinations}
\alias{setup_combinations}
\title{Set up combinations of parameters, options, climate and soil for mrunLWFB90}
\usage{
setup_combinations(param_len, options_len, soil_len, clim_len,
  all_combinations = F)
}
\arguments{
\item{param_len}{number of different parameters sets}

\item{options_len}{number of different option sets}

\item{soil_len}{number of different soil profiles}

\item{clim_len}{number of different climate data.tables}

\item{all_combinations}{set up all possible combinations? Defaults to FALSE.}
}
\value{
a data.frame with indexes of which object to combine with which object.
}
\description{
Set up combinations of parameters, options, climate and soil for mrunLWFB90
}
