% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slb1_meteo}
\alias{slb1_meteo}
\title{Meteorological Data from the Solling Beech and Spruce experimental site}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 19724 rows and 9 columns.}
\usage{
slb1_meteo
}
\description{
A dataset containing daily weather variables for the period 1960-2013
}
\keyword{datasets}
