% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slb1_prec2013_hh}
\alias{slb1_prec2013_hh}
\title{Hourly precipitation data from Solling Beech experimental site 'SLB1' for year 2013}
\format{
A data.frame with 8760 rows and 2 variables
\describe{
  \item{dates}{date}
  \item{prec}{hourly sum of precipitation, mm}
}
}
\usage{
slb1_prec2013_hh
}
\description{
Hourly precipitation data from Solling Beech experimental site 'SLB1' for year 2013
}
\keyword{datasets}
