% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slb1_standprop}
\alias{slb1_standprop}
\title{Annual stand properties of the Solling Beech experimental site 'SLB1'}
\format{A data.table with 49 rows and 7 variables
\describe{
  \item{year}{Year of observation}
  \item{species}{Tree species}
  \item{age}{age of the main stand}
  \item{height}{average height of the trees, m}
  \item{maxlai}{maximum leaf area index, m2 m-2}
  \item{sai}{stem area index, m2 m-2 }
  \item{densef}{stand density}
  \item{clay}{clay content, mass-\% }
  \item{c_org}{organic carbon content, mass-\%}
}}
\usage{
slb1_standprop
}
\description{
A dataset containing the forests's stand properties
}
\keyword{datasets}
