% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_to_param.R
\name{soil_to_param}
\alias{soil_to_param}
\title{Split up soil into materials and soil nodes.}
\usage{
soil_to_param(soil)
}
\arguments{
\item{soil}{data.frame with soil soil depths ('upper', 'lower') and
hydraulic parameters ('ths', 'thr', 'alpha', 'npar', 'ksat', 'tort', 'gravel')}
}
\value{
a list with data.table objects 'soil_nodes' and 'soil_materials'
}
\description{
Split up soil into materials and soil nodes.
}
