% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standprop_yrly_to_param.R
\name{standprop_yearly_to_param}
\alias{standprop_yearly_to_param}
\title{Transfer standproperties height, maxlai, sai, densef, age to parameter-list}
\usage{
standprop_yearly_to_param(standprop_yearly, param.b90, out.years)
}
\arguments{
\item{standprop_yearly}{a data.frame or data.table with columns 'year', 'height', 'maxlai', 'sai', 'densef', 'age'}

\item{out.years}{}
}
\value{
the param.b90 list-object with updated items maxlai, height, height.ini,
sai, sai.ini, densef, densef.ini, age, age.ini.
}
\description{
Takes a data.frame of yearly standproperties, trims/extends the columns height, maxlai,
sai, densef, and age for the years in out.years, and updates
}
