% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStand.R
\name{make_standprop}
\alias{make_standprop}
\title{Create the daily standproperties from parameters}
\usage{
make_standprop(options.b90, param.b90, out.years)
}
\arguments{
\item{options.b90}{a list of model control options}

\item{param.b90}{a parameter list-object}

\item{out.years}{years for which values are returned.}
}
\value{
a data.table containing daily sequences of 'age', 'height', 'sai', 'densef', and 'lai'.
}
\description{
Creates daily sequences of 'age', 'height', 'sai', 'densef', and 'lai' from parameter and options
using \code{\link{approx_standprop}} and \code{\link{MakeSeasLAI}}.
}
