% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSeasLAI.R
\name{MakeSeasLAI}
\alias{MakeSeasLAI}
\title{Construct the seasonal course of leaf area index from parameters}
\usage{
MakeSeasLAI(
  method = "b90",
  year,
  maxlai,
  winlaifrac = 0,
  budburst.doy = 121,
  leaffall.doy = 279,
  emerge.dur = 28,
  leaffall.dur = 58,
  shape.optdoy = 220,
  shape.budburst = 0.5,
  shape.leaffall = 10,
  lai.doy = c(1, 121, 150, 280, 320, 365),
  lai.frac = c(0, 0, 0.5, 1, 0.5, 0)
)
}
\arguments{
\item{method}{Name of method for generating the sequence. Must be one of "b90", "linear", "Coupmodel".}

\item{year}{Vector of years to be returned.}

\item{maxlai}{Maximum leaf are index.}

\item{winlaifrac}{Fraction of \code{maxlai} during winter (ignored when method = 'linear').}

\item{budburst.doy}{Budburst day of year (ignored when method = 'linear').}

\item{leaffall.doy}{Day of year when leaf fall begins (ignored when method = 'linear').}

\item{emerge.dur}{Number of days from budburst until maximum leaf area index is reached.}

\item{leaffall.dur}{Number of days until minimum leaf are index is reached.}

\item{shape.optdoy}{Day of year when optimum value is reached (required when method = "Coupmodel").}

\item{shape.budburst}{Shape parameter for the growth phase (required when method = "Coupmodel").}

\item{shape.leaffall}{Shape parameter growth cessation (required when method = "Coupmodel").}

\item{lai.doy}{Integer vector of days of years.}

\item{lai.frac}{Vector of values of fractional leaf area index corresponding
to lai.doy (required when method = "linear").}
}
\value{
A vector of daily lai values covering the years specified.
}
\description{
A daily sequence of leaf area index is derived from maximum and minimum values,
dates and shape parameters using different methods.
}
\examples{

# lai courses
lai_b90 <- MakeSeasLAI(method = "b90",
                       year = 2001,
                       maxlai = 5,
                       winlaifrac = 0,
                       budburst.doy = 121,
                       leaffall.doy = 280,
                       emerge.dur = 15,
                       leaffall.dur = 30)

lai.doy <- c(1,110,117,135,175,220,250,290,365)
lai.frac <- c(0.1,0.1,0.5,0.7,1.2,1.2,1.0,0.1,0.1)
lai_linear <- MakeSeasLAI(method = "linear",
                          year = 2001,
                          maxlai = 5,
                          lai.doy = lai.doy,
                          lai.frac = lai.frac)

lai_coupmodel <- MakeSeasLAI(method = "Coupmodel",
                             year = 2001,
                             maxlai = 5,
                             winlaifrac = 0.1,
                             budburst.doy = 110,
                             leaffall.doy = 280,
                             shape.optdoy = 180,
                             shape.budburst = 0.5,
                             shape.leaffall = 5)

plot(lai_b90, type = "n", xlab = "doy", ylab = "lai [m²/m²]", ylim = c(0,6))
lines(lai_b90, col ="green",lwd = 2,)
lines(lai_linear, col ="red",lwd = 2)
lines(lai_coupmodel, col ="blue",lwd = 2)

# incorparating between-year variability
years <- 2001:2003
lai <- MakeSeasLAI(method = "Coupmodel",
                   year = years,
                   maxlai = c(4,6,5),
                   budburst.doy = c(100,135,121),
                   leaffall.doy = 280,
                   shape.budburst = c(3,1,0.3),
                   shape.leaffall = 3,
                   shape.optdoy =c(210,180,240) )

dates <- seq.Date(as.Date("2001-01-01"), as.Date("2003-12-31"), by = "day")
plot(dates,lai, col = "green", ylab = "lai [m²/m²]",
     type ="l", xlab = "", lwd = 2)

}
