% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msiterunLWFB90.R
\name{msiterunLWFB90}
\alias{msiterunLWFB90}
\title{Make a multi-site simulation using list of inputs.}
\usage{
msiterunLWFB90(param.b90, options.b90, climate, soil = NULL,
  all_combinations = F, multirun.dir = "MultiRuns/",
  keep.subdirs = FALSE, cores = 3, showProgress = TRUE, ...)
}
\arguments{
\item{param.b90}{Named list of parameters to be used in all simulations, or list of lists with different parameter sets.}

\item{options.b90}{Named list of model control option to be used in all simulations, or list of lists with different options sets.}

\item{climate}{data.frame with climate data passed to \code{\link[runLWFB90]{runLWFB90}}, or a list of climate data.frames}

\item{soil}{data.frame with soil properties passed to \code{\link[runLWFB90]{runLWFB90}}, or a list of lists with different soil profiles}

\item{all_combinations}{logical wether to set up all possible combinations of
param.b90, options, climate and soil lists. See details.}

\item{multirun.dir}{the directory where to create the subdirectories for the the single runs. Default 'MultiRuns/'}

\item{keep.subdirs}{keep sub-directories of the single runs? Default is FALSE}

\item{cores}{number of cores to use for parallel processing}

\item{showProgress}{Show progressbar?}

\item{...}{Further arguments passed to \code{\link[runLWFB90]{runLWFB90}}.}
}
\value{
A named list with the results for the single runs. Simulation errors are passed on.
}
\description{
Repeatedly calls \code{\link[runLWFB90]{runLWFB90}} for lists of climate data, soil data,
parameter and options lists, stores the single run results in subdirectories and returns the results as a named list.
}
