% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantb90.R
\name{plant.b90}
\alias{plant.b90}
\title{Interpolate plant properties using the 'b90' method.}
\usage{
plant.b90(minval, maxval, doy.incr, incr.dur, doy.decr, decr.dur, maxdoy)
}
\arguments{
\item{minval}{minimum value}

\item{maxval}{maximum value}

\item{doy.incr}{date (day of year) when increasing to maxval begins}

\item{incr.dur}{duration (number of days) since doy.incr until maxval is rached}

\item{doy.decr}{date (day of year) when decreasing to minval begins}

\item{decr.dur}{duration (number of days) since doy.decr until minval is rached}

\item{maxdoy}{length of the year, 366 for leap years, 365 for normal years}
}
\value{
numeric vector of length \code{maxdoy}
}
\description{
Creates a daily sequence for one year from parameters
}
\examples{
plot(plant.b90(minval = 0,maxval=1,
               doy.incr = 121,incr.dur = 28,
               doy.decr = 280, decr.dur = 50,
               maxdoy = 365))
}
