% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantcoupmodel.R
\name{plant.coupmodel}
\alias{plant.coupmodel}
\title{Interpolate plant properties using the 'Coupmodel' method.}
\usage{
plant.coupmodel(minval, maxval, doy.incr, doy.max, doy.min, shape.incr,
  shape.decr, maxdoy)
}
\arguments{
\item{minval}{minimum value}

\item{maxval}{maximum value}

\item{doy.incr}{date (day of year) when increasing to maxval begins}

\item{doy.max}{date (day of year) when maxval is reached}

\item{doy.min}{date (day of year) when minval is reached}

\item{shape.incr}{shape parameter of the increasing phase}

\item{shape.decr}{shape parameter of the decreasing phase}

\item{maxdoy}{length of the year, 366 for leap years, 365 for normal years}
}
\value{
numeric vector of length \code{maxdoy}
}
\description{
Creates a daily sequence for one year from parameters
}
\examples{
plot(plant.coupmodel(0,5, 121, 200, 280, 0.3, 3, 365))
}
