% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richter_prec_corr.R
\name{prec_corr}
\alias{prec_corr}
\title{Correct rain gauge precipitation data for wind and evaporation errors after Richter (1995)}
\usage{
prec_corr(dates, tavg, prec, station.exposure = "mg",
  full.output = FALSE)
}
\arguments{
\item{dates}{a vector of Dates}

\item{tavg}{a vector of air temperature}

\item{prec}{a vector of rainfall}

\item{station.exposure}{situation of the weather station where prec was measured: one of 'frei', 'lg', 'mg', 'sg'
(corresponding to full exposure, low protected, moderate protected, strong protected situation)}

\item{full.output}{logical wether to return the full data set additionally including input data, correction coefficients.}
}
\value{
a vector of corrected rainfall data, or (if full.output == TRUE) a data.table containing the input objects, the month,
the precipitation type ('N4So': liquid rain, summer; 'N4Wi' liquid rain, winter; 'N8' = sleet, 'N7' = snow),
correction coefficients epsilon and b, and the corrected rainfall.
}
\description{
Correct rain gauge precipitation data for wind and evaporation errors after Richter (1995)
}
\examples{

prec_corr(slb1_meteo$dates, slb1_meteo$tmean, slb1_meteo$prec)
prec <- with(slb1_meteo[year(dates) \%in\% 2001:2010,],
             data.frame(dates = dates, meas = prec,
                        frei = prec_corr(dates, tmean, prec, station.exposure = "frei"),
                        lg = prec_corr(dates, tmean, prec, station.exposure = "lg"),
                        mg = prec_corr(dates, tmean, prec, station.exposure = "mg"),
                        sg = prec_corr(dates, tmean, prec, station.exposure = "sg"))
)


plot(prec$dates, cumsum(prec$frei), type = "l", col = "violet", xlab = "dates", ylab = "cum. precipitation (mm)")
lines(prec$dates, cumsum(prec$lg), col = "blue")
lines(prec$dates, cumsum(prec$mg), col = "green")
lines(prec$dates, cumsum(prec$sg), col = "red")
lines(prec$dates, cumsum(prec$meas))
legend('bottomright', c('frei', "lg", "mg", "sg"), col = c("violet", "blue", "green", "red", "black"), lty = 1, pch = NULL )
}
\references{
Richter, D. (1995): "Ergebnisse methodischer Untersuchungen
\emph{Berichte des Deutschen Wetterdienstes}, \bold{194}, 93 pp, Offenbach, Germany
}
