% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_lwfbrook90.R
\name{r_lwfbrook90}
\alias{r_lwfbrook90}
\title{Define the function to run LWFBrook90 in R}
\usage{
r_lwfbrook90(siteparam, climveg, param, pdur, soil_materials, soil_nodes,
  precdat = NULL, output, output_log = TRUE)
}
\arguments{
\item{siteparam}{a [1,6] matrix with site level information: start year, start doy,
latitude, initial snow, initial groundwater, precipitation interval}

\item{climveg}{a matrix with 15 columns of climatic and vegetation data: year, month, day
globrad (MJ m-2 d-1), tmax (degC), tmin (degC), vappres (kPar), wind (m s-1), prec (mm), mesfl (mm),
densef (-), stand height (m), lai (m2 m-2), sai (m2 m-2), age (years)}

\item{param}{a vector of model input parameters}

\item{pdur}{a [1,12]-matrix of precipitation durations (hours) for each monath}

\item{soil_materials}{a matrix of the 8 soil materials parameters. When imodel = 1 (Mualem-van Genuchten), these refer to:
mat, ths, thr , alpha (m-1), npar, ksat (mm d-1), tort (-), stonef (-). When imodel = 2 (Clapp-Hornberger):
mat, thsat, thetaf, psif (kPa), bexp, kf (mm d-1), wetinf (-), stonef (-)}

\item{soil_nodes}{a matrix of the 6 soil model layers with columns nl (layer number),
layer midpoint (m), thickness (mm), mat, psiini (kPa), rootden (-).}

\item{precdat}{a matrix precipitation interval data with 6 columns:
year, month, day, interval-number (1:precint), prec, meslflp}

\item{output}{a [12,5] matrix of output selection settings}

\item{output_log}{logical wether to write the output-logfile 'Log.txt'. This is where the
commandline-feed of the original fortan program is written.}
}
\description{
Define the function to run LWFBrook90 in R
}
\details{
model output is written to comma-separated files (.ASC) in the working directory.
}
