% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLWFB90.R
\name{runLWFB90}
\alias{runLWFB90}
\title{Run the LWF-Brook90 hydrological model and return results}
\usage{
runLWFB90(project.dir, options.b90, param.b90, climate, precip = NULL,
  soil = NULL, outputmat = setoutput_LWFB90(),
  output.param.options = TRUE, run = TRUE, read.output = TRUE,
  output.log = TRUE, verbose = TRUE)
}
\arguments{
\item{project.dir}{directory-name of the project to which output files
are written. Will be created, if not existing.}

\item{options.b90}{named list of model control options. Use
\code{\link{setoptions_LWFB90}} to generate a list with default model control options.}

\item{param.b90}{Named list of model input parameters. Use
\code{\link{setparam_LWFB90}} to generate a list with default model parameters.}

\item{climate}{data.frame with daily climate data. The names of climate have to
correspond to arguments \emph{dates}, \emph{tmax}, \emph{tmin}, \emph{wind}, \emph{prec}, \emph{vappres},
\emph{globrad}, \emph{sunhours}) of \code{\link{writeClimate.in}}.}

\item{precip}{data.frame to supply precipitation data separately from climate data.
Can be used to provide sub-day resolution precipitation data to LWFBrook90.}

\item{soil}{data.frame containing the hydraulic properties of the soil layers.
Each row represents one layer, containing the layers' boundaries and soil hydraulic parameters.
The columns names for the upper and lower layer boundaries are \emph{upper} and \emph{lower} (m, negative downwards),
the parameters of the van Genuchten retention functions are \emph{ths}, \emph{thr},
 \emph{alpha} [m-1], \emph{npar}, and the parameters of the Mualem conductivity function
 \emph{ksat} [mm d-1] and \emph{tort}. The volume fraction of stones has to be named \emph{gravel}.
 If the soil data.frame is not provided, list items soil_nodes and soil_materials of param.b90 are used for the simulation.
 These have to be set up in advance.}

\item{outputmat}{a [10,5]-matrix flagging the desired model-output. Use
\code{\link{setoutput_LWFB90}} to generate and edit a default output matrix.}

\item{output.param.options}{append 'param.b90', 'options.b90', 'soil' and daily plant
properties ('plant.devt', as derived from parameters and written to 'climate.in') to the result?}

\item{run}{run LWF-Brook90 or only return model input objects?
Useful to inspect the effects of options and parameters on model input. Default is TRUE.}

\item{read.output}{read and return simulation results from out.dir? Default is TRUE.}

\item{output.log}{write the logfile Log.txt? Default is TRUE.}

\item{verbose}{print messages to the console? Default is TRUE.}
}
\value{
Returns the model-output from the files found in 'project.dir' as a list of data.tables,
along with the execution time of the simulation, and model input if desired.
}
\description{
Sets up the input objects for LWF-Brook90 hydrological model, starts the model,
and returns the results.
}
\examples{

#Set up lists containing model control options and model parameters:

param.b90 <- setparam_LWFB90()
options.b90 <- setoptions_LWFB90()

# Set start and end Dates for the simulation

options.b90$startdate <- as.Date("2000-01-01")
options.b90$enddate <- as.Date("2004-12-31")

# Derive soil hydraulic properties from soil physical properties
# using pedotransfer functions

soil <- cbind(soil_slb1, hydpar_wessolek_mvg(soil_slb1$texture))

# Run LWF-Brook90
b90.result <- runLWFB90(project.dir = "example_run_b90",
                      options.b90 = options.b90,
                      param.b90 = param.b90.b90,
                      climate = meteo_slb1,
                      soil = soil)
}
