% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLWFB90.R
\name{runLWFB90}
\alias{runLWFB90}
\title{Run the LWF-Brook90 hydrological model}
\usage{
runLWFB90(project.dir = "runLWFB90/", options.b90, param.b90, climate,
  precip = NULL, soil = NULL, output = setoutput_LWFB90(),
  output_fun = NULL, rtrn.input = TRUE, rtrn.output = TRUE,
  read.output = TRUE, chk.input = TRUE, output.log = TRUE,
  run = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{project.dir}{directory-name of the project to which output files
are written. Will be created, if not existing. Defaults to 'runLWFB90/'}

\item{options.b90}{named list of model control options. Use
\code{\link{setoptions_LWFB90}} to generate a list with default model control options.}

\item{param.b90}{Named list of model input parameters. Use
\code{\link{setparam_LWFB90}} to generate a list with default model parameters.}

\item{climate}{data.frame with daily climate data. See details for the required variables.}

\item{precip}{data.frame with columns 'dates' and 'prec' to supply precipitation data separately from climate data.
Can be used to provide sub-day resolution precipitation data to LWFBrook90. For each day in dates,
1 (daily resolution) to 240 values of precipitation can be provided, with the number of values
per day defined in \code{options.b90$prec.interval}.}

\item{soil}{data.frame containing the hydraulic properties of the soil layers. See details.}

\item{output}{a [10,5]-matrix flagging the desired model-output. Use
\code{\link{setoutput_LWFB90}} to generate and edit a default output matrix.}

\item{output_fun}{a function or a list of functions to be performed on the output objects selected by \code{output}.
Can be used to aggregate output or calculate goodness of fit measures.
Disabled when read.output = FALSE.}

\item{rtrn.input}{append 'param.b90', 'options.b90', 'soil' and daily plant
properties ('standprop_daily', as derived from parameters) to the result?}

\item{rtrn.output}{return the simulation results? Disabled when read.output = FALSE.}

\item{read.output}{read the simulation result files from project.dir? Default is TRUE.}

\item{chk.input}{logical wether to check param.b90, options.b90, climate, precip, soil, and obs
for completeness and consistency.}

\item{output.log}{write the logfile 'Log.txt' to the 'project.dir'? Default is TRUE.}

\item{run}{run LWF-Brook90 or only return model input objects?
Useful to inspect the effects of options and parameters on model input. Default is TRUE.}

\item{verbose}{print messages to the console? Default is TRUE.}

\item{...}{additional arguments passed to \code{output_fun}.}
}
\value{
A list containing the model input (except for climate), the contents of
the LWF-Brook90 output files found in 'project.dir', and the return values of \code{output_fun}.
}
\description{
Sets up the input objects for the LWF-Brook90 hydrological model, starts the model,
and returns the results if desired.
}
\section{Climate data}{

The \code{climate} data.frame must contain the following variable in columns named
\code{dates} (Date), \code{tmax} (deg C), \code{tmin} (deg C), \code{tmean} (deg C),
\code{wind} (m s-1), \code{prec} (mm) , \code{vappres} (kPa), and either \code{globrad} (MJ d-1 m-2)
or \code{sunhours} (hours). When using \code{sunhours}, please set \code{options.b90$fornetrad = 'sunhours'}.
}

\section{Soil parameters}{

Each row of \code{soil} represents one layer, containing the layers' boundaries and soil hydraulic parameters.
The column names for the upper and lower layer boundaries are \code{upper} and \code{lower} (m, negative downwards).
When using options.b90$imodel = 'MvG', the hydraulic parameters are  \code{ths}, \code{thr},
 \code{alpha} (m-1), \code{npar}, \code{ksat} (mm d-1) and \code{tort}.  With options.b90$imodel = 'CH',
 the parameters are \code{thsat}, \code{thetaf}, \code{psif}(kPa), \code{bexp},
 \code{kf} (mm d-1), and \code{wetinf}. For both parameterizations, the volume fraction of stones has to be named \code{gravel}.
 If the soil data.frame is not provided, list items 'soil_nodes' and 'soil_materials'
 of param.b90 are used for the simulation. These have to be set up in advance, see \code{\link{soil_to_param}}.
}

\examples{
# Set up lists containing model control options and model parameters:
param.b90 <- setparam_LWFB90()
options.b90 <- setoptions_LWFB90()

# Set start and end Dates for the simulation
options.b90$startdate <- as.Date("2002-2-15")
options.b90$enddate <- as.Date("2005-07-5")

# Derive soil hydraulic properties from soil physical properties
# using pedotransfer functions
soil <- cbind(slb1_soil, hydpar_wessolek_mvg(slb1_soil$texture))

# Run LWF-Brook90
b90.result <- runLWFB90(project.dir = "example_run_b90",
                        options.b90 = options.b90,
                        param.b90 = param.b90,
                        climate = slb1_meteo,
                        soil = soil)

# use a function to be performed on the output:
# aggregate soil water storage down to a specific layer
agg_swat <- function(x, layer) {
  out <- aggregate(SWATI~YR+DOY,
                   x$SWATDAY.ASC,
                   FUN = sum,
                   subset = NL <= layer)
  out[order(out$YR, out$DOY),]}

# run model, without returning the original output.
b90.aggswat <- runLWFB90(project.dir = "example_run_b90",
                         options.b90 = options.b90,
                         param.b90 = param.b90,
                         climate = slb1_meteo,
                         soil = soil,
                         output_fun = list(swat = agg_swat),
                         rtrn.output = F,
                         layer = 10) #' passed to output_fun
str(b90.aggswat$output_fun$swat)

}
