% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLWFB90.R
\name{runLWFB90}
\alias{runLWFB90}
\title{Run the LWF-Brook90 hydrological model and return results}
\usage{
runLWFB90(project.dir, options.b90, param.b90, climate, soil = NULL,
  outputmat = setoutput_LWFB90(), keep.log.on.success = T,
  out.dir = "out/", output.param.options = TRUE,
  keep.outputfiles = TRUE, verbose = TRUE)
}
\arguments{
\item{options.b90}{named list of model control options. Use
\code{\link{MakeOptions.B90}} to generate a list with default model control options.}

\item{param.b90}{Named list of model input parameters. Use
\code{\link{MakeParam.B90}} to generate a list with default model parameters.}

\item{climate}{data.frame with daily climate data. The names of climate have to
correspond to arguments \emph{dates}, \emph{tmax}, \emph{tmin}, \emph{wind}, \emph{prec}, \emph{vappres},
\emph{globrad}, \emph{sunhours}) of \code{\link{writeClimate.in}}.}

\item{soil}{data.frame containing the hydraulic properties of the soil layers.
Each row represents one layer, containing the layers' boundaries and soil hydraulic parameters.
The columns names for the upper and lower layer boundaries are \emph{upper} and \emph{lower} (m, negative downwards),
the parameters of the van Genuchten retention functions are \emph{ths}, \emph{thr},
 \emph{alpha} [m-1], \emph{npar}, and the parameters of the Mualem conductivity function
 \emph{ksat} [mm d-1] and \emph{tort}. The volume fraction of stones has to be named \emph{gravel}.
 If the soil data.frame is not provided, list items soil_nodes and soil_materials of param.b90 are used for the simulation.
 These have to be set up in advance.}

\item{outputmat}{a [10,5]-matrix flagging the desired model-output. Use
\code{\link{choose_output.B90}} to generate and edit default output matrix.}

\item{keep.log.on.success}{keep the file 'output.log' after a successful simulation?
In case of simulation errors the 'output.log' file (if specified) is kept anyway for inspection purposes.}

\item{out.dir}{path where to write the output-files.}

\item{output.param.options}{append 'param.b90', 'options.b90', 'soil' and daily plant
properties ('plant.devt', as derived from parameters and written to 'climate.in') to the result?}

\item{keep.outputfiles}{keep the model .asc output files after running and
returning the output?}

\item{verbose}{print messages to the console? Default is TRUE.}

\item{run.model}{run the model or only return model input objects? Default is TRUE.}
}
\value{
Returns the model-output from the files found in 'out.dir' as a list of data.tables,
along with the execution time of the simulation, and input-parameters and options if desired.
}
\description{
Takes all necessary information needed to run the LWF-Brook90 hydrological model,
writes input files, starts the external executable via a system-call and returns
the results.
#' @param project.dir directory-name of the project where the input and output files
are located. Will be created, if not existing.
}
\examples{

#Set up lists containing model control options and model parameters:

param.b90 <- MakeParam.B90()
options.b90 <- MakeOptions.B90()

# Set start and end Dates for the simulation

options.b90$startdate <- as.Date("2000-01-01")
options.b90$enddate <- as.Date("2004-12-31")

# Derive soil hydraulic properties from soil physical properties
# using pedotransfer functions

soil <- cbind(soil_slb1, hydpar_wessolek_mvg(soil_slb1$texture))

# Run LWF-Brook90
b90.result <- Run.B90(project.dir = "example_run_b90",
                      options.b90 = options.b90,
                      param.b90 = param.b90.b90,
                      climate = meteo_slb1,
                      soil = soil,
                      path_b90.exe = "b90.exe")
}
