% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setoptionsLWFB90.R
\name{setoptions_LWFB90}
\alias{setoptions_LWFB90}
\title{Create a list of model control options}
\usage{
setoptions_LWFB90(...)
}
\arguments{
\item{...}{named values to be included in return value}
}
\value{
Returns a list of default model control options for use as 'options.b90'-argument in \code{\link{runLWFB90}}.
}
\description{
Create a list of model control options
}
\examples{
# Default options
options.b90 <- setoptions_LWFB90()
# Include specific options
options.b90_dynamic_phenology <- setoptions_LWFB90(budburst.method = 'Menzel',
leaffall.method ='vonWilpert')
}
