% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setoutput_LWFB90.R
\name{setoutput_LWFB90}
\alias{setoutput_LWFB90}
\title{Select output for LWF-Brook90}
\usage{
setoutput_LWFB90(outputmat = NULL, edit = FALSE)
}
\arguments{
\item{outputmat}{optional [5,10]-matrix, opened for editing.
If no matrix is passed, a default selection of output values is opened in R's data-editor or returned
instantly, if edit=FALSE.}

\item{edit}{open R's data-editor ?}
}
\value{
a [5,10]-matrix containing 0 and 1 for use as 'outputmat'-argument in \code{\link{Run.B90}}
}
\description{
Opens the data editor with a [5,10] matrix for selecting the LWF-Brook90-output
by flagging cells with 1. Assign a name to the returned matrix and use it as
'outputmat'-argument in \code{\link{Run.B90}}. A default output-matrix is returned
without opening the data-editor when edit = FALSE.
}
\examples{

# create matrix with default selection
outmat <- choose_output.B90()

outmat[6,] <- 1

outmat_new <- setoutput_LWFB90(outmat)

outmat <- setoutput_LWFB90(edit = T)

}
