% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slb1_meteo}
\alias{slb1_meteo}
\title{Meteorological Data from the Solling Beech and Spruce experimental site}
\format{A data.frame with 19724 rows and 9 variables
\describe{
  \item{dates}{date}
  \item{tmin}{daily minimum temperature, degC}
  \item{tmax}{daily maximum temperature, degC}
  \item{tmean}{daily mean temperature, degC}
  \item{prec}{daily sum of precipitation, mm}
  \item{relhum}{relative Humidity, \%}
  \item{globrad}{daily sum of global radiation, MJ m-2}
  \item{wind}{daily mean wind speed measured at 10 m above ground, m s-2}
  \item{vappres}{daily vapour pressure, kPa}
}}
\usage{
slb1_meteo
}
\description{
A dataset containing daily weather variables for the period 1960-2013
}
\keyword{datasets}
